requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
charset-normalizer==3.3.2
coverage==7.6.0
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
peewee==3.15.4
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.3.1
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.8
python-dotenv==1.0.1
yarl==1.9.4
